+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.9.0-beta.2] - 2026-04-01
+
+### Changed
+
+- Print-friendly Markdown exports can now insert HTML page breaks between note sections, which also puts the table of contents on its own page when enabled
+- New print-friendly Markdown setting now controls whether page breaks replace section dividers in clipboard and new-note exports
+
 ## [1.9.0-beta.1] - 2026-03-15
 
 ### Changed